[EAK-365] base injector name has been shortened (BaseInjectorTemplateMethod -> BaseInjector)
